requirements.txt
fastapi
uvicorn[standard]

langgraph
langchain
langchain-openai
langchain-community
langchain_pinecone

unstructured
pydantic
python-dotenv
httpx
python-multipart

pinecone
motor
pymongo